Just the lexer and parser defintion files
These are just the lexer and parser definition files for parsing
jason.  They are not 100% correct due to my own misunderstanding
of what syntactailly correct json is.

I know that it is not 100% valid json as I got tired of typing
and reading 100% valid json.  At this point I just don't remmember
where I relaxed things a bit.

I also need to add an c/c++ module to act as an entry point
to this library in order to hide the parser.h and lexer.h
that are generated for and specific to this project.